SriLankanConstitutionAI/requirements.txt
gradio
langchain
langchain_community
ollama
pypdf
chromadb
requests
huggingface-hub
python-dotenv
langchain-core
PyPDF2
chroma
pydantic
gpt4all